Mon, Jun 10, 2024, 9:43 AM -03:00

+++ set1/meal/meal.py
@@ -9,7 +9,7 @@ def main():
 def convert(time):
     hours, minutes = time.split(":")
 
-    if int(hours) >= 7 and int(hours) <= 9:
+    if float(hours) >= 7 and float(hours) <= 9:
         print('breakfast time')
     elif int(hours) >= 12 and int(hours) <= 13:
         print('lunch time')
